requirements.txt
openai
python-dotenv
requests
prompt_toolkit